Divvy HorizontalLinkBar up into smaller pieces
