+++ src/atoms/LinkBarNav.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+const LinkBarNav = styled.nav`
+	background: ${props => props.background};
+	position: relative;
+	width: ${props => props.width};
+`;
+
+export { LinkBarNav };
